requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
pip==24.3.1
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.9.0
django-filter==25.2